Add PyPi package, add standard config paths
