--- liteshort.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python3
-
-from liteshort.main import app
-
-if __name__ == "__main__":
-    app.run()
