+++ setup/liteshort.ini
